simplify config
